Introduce gaussian PSF evaluate
