+++ gammapy/irf/psf/core.py
@@ -131,3 +131,85 @@ class ParametricPSF(IRF):
                         "".format(100 * fraction, theta, energy, radius)
                     )
         return ss
+
+    def plot_containment(self, fraction=0.68, ax=None, add_cbar=True, **kwargs):
+        """
+        Plot containment image with energy and theta axes.
+
+        Parameters
+        ----------
+        fraction : float
+            Containment fraction between 0 and 1.
+        add_cbar : bool
+            Add a colorbar
+        """
+        import matplotlib.pyplot as plt
+
+        ax = plt.gca() if ax is None else ax
+
+        energy = self.axes["energy_true"].center
+        offset = self.axes["offset"].center
+
+        # Set up and compute data
+        containment = self.containment_radius(energy, offset, fraction)
+
+        # plotting defaults
+        kwargs.setdefault("cmap", "GnBu")
+        kwargs.setdefault("vmin", np.nanmin(containment.value))
+        kwargs.setdefault("vmax", np.nanmax(containment.value))
+
+        # Plotting
+        x = energy.value
+        y = offset.value
+        caxes = ax.pcolormesh(x, y, containment.value, **kwargs)
+
+        # Axes labels and ticks, colobar
+        ax.semilogx()
+        ax.set_ylabel(f"Offset ({offset.unit})")
+        ax.set_xlabel(f"Energy ({energy.unit})")
+        ax.set_xlim(x.min(), x.max())
+        ax.set_ylim(y.min(), y.max())
+
+        if add_cbar:
+            label = f"Containment radius R{100 * fraction:.0f} ({containment.unit})"
+            ax.figure.colorbar(caxes, ax=ax, label=label)
+
+        return ax
+
+    def plot_containment_vs_energy(
+        self, fractions=[0.68, 0.95], thetas=[0, 1] * u.deg, ax=None, **kwargs
+    ):
+        """Plot containment fraction as a function of energy.
+        """
+        import matplotlib.pyplot as plt
+
+        ax = plt.gca() if ax is None else ax
+
+        energy = self.axes["energy_true"].center
+
+        for theta in thetas:
+            for fraction in fractions:
+                radius = self.containment_radius(energy, theta, fraction).squeeze()
+                kwargs.setdefault("label", f"{theta.to_value('deg')} deg, {100 * fraction:.1f}%")
+                ax.plot(energy.value, radius.value, **kwargs)
+
+        ax.semilogx()
+        ax.legend(loc="best")
+        ax.set_xlabel("Energy (TeV)")
+        ax.set_ylabel("Containment radius (deg)")
+
+    def peek(self, figsize=(15, 5)):
+        """Quick-look summary plots."""
+        import matplotlib.pyplot as plt
+
+        fig, axes = plt.subplots(nrows=1, ncols=3, figsize=figsize)
+
+        self.plot_containment(fraction=0.68, ax=axes[0])
+        self.plot_containment(fraction=0.95, ax=axes[1])
+        self.plot_containment_vs_energy(ax=axes[2])
+
+        # TODO: implement this plot
+        # psf = self.psf_at_energy_and_theta(energy='1 TeV', theta='1 deg')
+        # psf.plot_components(ax=axes[2])
+
+        plt.tight_layout()
